Add Makefile for linting and serving
The Makefile contains targets for linting for serving the website
locally.
This commit also adds a hack script for running pre-commit similar to
markdownlint, shellcheck and spellcheck. Now all of them can be run with
a simple `make lint` as long as podman or docker is available.

The README.md is also update to make use of the make serve command. The
image that was used in the readme no longer works with the versions we
are using.

Signed-off-by: Lennart Jern <lennart.jern@est.tech>