fix spellcheck glob
Since we're running the check in sh, it does not have globstar option
to handle '**/*.md' properly. It only checked exactly one level deep
markdown files. Spellcheck handles globs itself, so just adding quotes
is enough to fix it.

In this repo, it only ignored top-level ones, which luckily do not have
any spellcheck issues, so it is enough to just fix the glob.

Signed-off-by: Tuomo Tanskanen <tuomo.tanskanen@est.tech>